# apps/backend/api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Arenda Planform API
  version: 0.1.0
servers:
  - url: http://localhost:8080
paths:
...
  /recurring-operations:
    get:
      operationId: listRecurringOperations
      tags: [Recurring Operations]
      security:
        - sessionCookie: []
      responses:
        '200':
          description: Recurring operations list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecurringOperationsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /recurring-operations/{id}:
    get:
      operationId: getRecurringOperation
      tags: [Recurring Operations]
      security:
        - sessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Recurring operation details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecurringOperationResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      operationId: updateRecurringOperation
      tags: [Recurring Operations]
      security:
        - sessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecurringOperationUpdateRequest'
      responses:
        '200':
          description: Recurring operation updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecurringOperationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '403':
          $ref: '#/components/responses/SubscriptionBlocked'
    delete:
      operationId: deleteRecurringOperation
      tags: [Recurring Operations]
      security:
        - sessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Recurring operation deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /recurring-operations/{id}/pause:
    post:
      operationId: pauseRecurringOperation
      tags: [Recurring Operations]
      security:
        - sessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Recurring operation paused
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecurringOperationResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '403':
          $ref: '#/components/responses/SubscriptionBlocked'
  /recurring-operations/{id}/resume:
    post:
      operationId: resumeRecurringOperation
      tags: [Recurring Operations]
      security:
        - sessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Recurring operation resumed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecurringOperationResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '403':
          $ref: '#/components/responses/SubscriptionBlocked'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session_id
      description: >
        The actual cookie name depends on the `cookieSecure` setting:
        `session_id` when insecure and `__Host-session_id` when secure.
...
  schemas:
...
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        requestId:
          type: string
        code:
          type: string
          description: >-
            Machine-readable error code for failures the client must
            distinguish from the generic HTTP status (e.g.
            `membership_suspended`).
        errors:
          type: array
          items:
            $ref: '#/components/schemas/ProblemError'
    ProblemError:
      type: object
      required: [field, detail]
      properties:
        field:
          type: string
          description: JSON key of the invalid field (e.g. "floor").
        detail:
          type: string
          description: Human-readable explanation of the validation failure.
...
    OperationType:
      type: string
      enum: [income, expense]
...
    RecurringOperationUpdateRequest:
      type: object
      properties:
        type:
          $ref: '#/components/schemas/OperationType'
        category_id:
          type: string
          format: uuid
        name:
          type: string
          maxLength: 50
        amount_kopecks:
          type: integer
          format: int64
          minimum: 0
        start_date:
          type: string
          format: date
        payment_day:
          type: integer
          minimum: 1
          maximum: 31
        end_date:
          type: string
          format: date
        comment:
          type: string
          maxLength: 2000
        periodicity:
          type: string
          enum: [monthly, yearly]
        apply_from_date:
          type: string
          format: date
        reminder_offset_days:
          type: integer
          nullable: true
          enum: [0, 1, 3, 7]
          description: 0 clears any existing reminders; 1/3/7 reschedules reminders for future generated operations.
    RecurringOperationResponse:
      type: object
      required: [id, owner_id, type, category_id, category_name, name, amount_kopecks, start_date, payment_day, periodicity, status, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        owner_id:
          type: string
          format: uuid
        property_id:
          type: string
          format: uuid
          nullable: true
          description: Null when the property was deleted in detach mode.
        lease_id:
          type: string
          format: uuid
          nullable: true
        type:
          $ref: '#/components/schemas/OperationType'
        category_id:
          type: string
          format: uuid
        category_name:
          type: string
        name:
          type: string
        amount_kopecks:
          type: integer
          format: int64
        start_date:
          type: string
          format: date
        payment_day:
          type: integer
        end_date:
          type: string
          format: date
          nullable: true
        periodicity:
          type: string
          enum: [monthly, yearly]
        status:
          type: string
          enum: [active, paused]
        comment:
          type: string
          nullable: true
        reminder_offset_days:
          type: integer
          nullable: true
          enum: [1, 3, 7]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    RecurringOperationsResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/RecurringOperationResponse'
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    SubscriptionBlocked:
      description: Subscription is blocked; renew to continue mutating data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
...